Refactor class names
- Change <type>NottificationRequest to Deliverable<type>Notification
- Move IngestAdjunctRequest back
