+++ src/protagonist/CleanupHandlerTests/AssetDeletedHandlerTests.cs
@@ -324,7 +324,7 @@ public class AssetDeletedHandlerTests
     public async Task Handle_InvalidatesImagePath_IfDeliveryChannels()
     {
         // Arrange
-        var cleanupRequest = new DeletedNotificationRequest<Asset>()
+        var cleanupRequest = new DeliverableDeletedNotification<Asset>()
         {
             Deliverable = new Asset()
             {
@@ -419,7 +419,7 @@ public class AssetDeletedHandlerTests
     public async Task Handle_InvalidatesImagePath_IfImageAssetFamily()
     {
         // Arrange
-        var cleanupRequest = new DeletedNotificationRequest<Asset>()
+        var cleanupRequest = new DeliverableDeletedNotification<Asset>()
         {
             Deliverable = new Asset()
             {
@@ -462,7 +462,7 @@ public class AssetDeletedHandlerTests
     public async Task Handle_DoesNotCreateInvalidation_IfFileAssetFamily()
     {
         // Arrange
-        var cleanupRequest = new DeletedNotificationRequest<Asset>()
+        var cleanupRequest = new DeliverableDeletedNotification<Asset>()
         {
             Deliverable = new Asset()
             {
@@ -501,7 +501,7 @@ public class AssetDeletedHandlerTests
     public async Task Handle_DoesNotCreateInvalidation_IfDeleteFromDoesNotContainCdn()
     {
         // Arrange
-        var cleanupRequest = new DeletedNotificationRequest<Asset>()
+        var cleanupRequest = new DeliverableDeletedNotification<Asset>()
         {
             Deliverable = new Asset()
             {
@@ -540,7 +540,7 @@ public class AssetDeletedHandlerTests
     public async Task Handle_ReturnsFalse_IfInvalidationFails()
     {
         // Arrange
-        var cleanupRequest = new DeletedNotificationRequest<Asset>()
+        var cleanupRequest = new DeliverableDeletedNotification<Asset>()
         {
             Deliverable = new Asset()
             {
@@ -608,7 +608,7 @@ public class AssetDeletedHandlerTests
     
     private QueueMessage CreateMinimalQueueMessage()
     {
-        var cleanupRequest = new DeletedNotificationRequest<Asset>
+        var cleanupRequest = new DeliverableDeletedNotification<Asset>
         {
             Deliverable = new Asset()
             {
