+++ src/protagonist/DLCS.Model/Messaging/DeliverableDeletedNotification.cs
@@ -3,7 +3,7 @@ using DLCS.Model.PathElements;
 
 namespace DLCS.Model.Messaging;
 
-public class DeletedNotificationRequest<T> where T : IDeliverable
+public class DeliverableDeletedNotification<T> where T : IDeliverable
 {
     public T? Deliverable { get; set; }
