+++ src/protagonist/Engine/Ingest/AdjunctIngester.cs
@@ -1,7 +1,6 @@
 ﻿using DLCS.Core.Types;
 using DLCS.Model.Customers;
 using DLCS.Model.Messaging;
-using DLCS.Model.Messaging.Adjunct;
 using Engine.Data;
 
 namespace Engine.Ingest;
